% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{chain}
\alias{chain}
\alias{chain.rtree}
\alias{chain.list}
\alias{chain.numeric}
\alias{chain.character}
\alias{c_}
\title{Chain constructor}
\usage{
chain(...)

\method{chain}{rtree}(...)

\method{chain}{list}(...)

\method{chain}{numeric}(...)

\method{chain}{character}(...)

c_(...)
}
\arguments{
\item{...}{characters or nodes.}
}
\value{
A chain.
}
\description{
The function \code{chain} creates an 'rtree' object which is a chain, 
i.e. a totally ordered tree.
}
