% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{children}
\alias{children}
\title{Children of a node}
\usage{
children(.node, .tree, degree = 1L)
}
\arguments{
\item{.node}{node or character. The node or node label considered.}

\item{.tree}{A tree.}

\item{degree}{integer. Currently not used.}
}
\value{
A (possibly empty) list of nodes.
}
\description{
The function \code{children} returns the children 
of a node in a given tree.
}
