% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaves.R
\name{cut_leaves}
\alias{cut_leaves}
\title{Cut the leaves of a tree}
\usage{
cut_leaves(.tree)
}
\arguments{
\item{.tree}{A tree.}
}
\value{
A tree.
}
\description{
The function \code{cut_leaves} cuts the leaves 
in \code{.tree}.
}
