% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descendants.R
\name{descendants}
\alias{descendants}
\alias{descendants.rtree}
\title{Descendants of a node}
\usage{
descendants(.node, .tree, include_node = FALSE)

\method{descendants}{rtree}(.node, .tree, include_node = FALSE)
}
\arguments{
\item{.node}{node or character. The node or node label considered.}

\item{.tree}{A tree.}

\item{include_node}{logical. Currently not used.}
}
\value{
A (possibly empty) list of nodes.
}
\description{
The function \code{descendants} returns the descendants 
of a node in a given tree.
}
