% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{\%->\%}
\alias{\%->\%}
\title{Add a tree at the bottom of a chain}
\usage{
e1 \%->\% e2
}
\arguments{
\item{e1}{A chain.}

\item{e2}{A rooted tree.}
}
\value{
A rooted tree.
}
\description{
The function \code{\%->\%} links the rooted tree \code{e2} 
to the bottom of the chain \code{e1}.
}
\examples{
## Chain
(tr0 = c_("Bob", "Carl", "Daniel"))

## Rooted tree
(tr1 = c_("Bill", "Caroline", "Dimitri", "Enoc"))
(tr2 = c_("John", "Thomas"))
(tr3 = r_("Alice", s = list(tr1, tr2))) 

## Linking both
tr0 \\\%->\\\% tr3

}
