% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.binary_tree.R
\name{is.binary_tree}
\alias{is.binary_tree}
\title{Test if a tree is a binary tree}
\usage{
is.binary_tree(x)
}
\arguments{
\item{x}{A tree to be tested.}
}
\value{
A logical.
}
\description{
The function \code{is.binary_tree} tests if 
a tree is binary.
}
\examples{
## FALSE
is.binary_tree(empty_tree())

## FALSE
tr = r_(s = list(r_("toto"), r_("tata")))
is.binary_tree(tr) # unrooted tree

## TRUE
tr = r_("titi", s = list(r_("toto"), r_("tata")))
is.binary_tree(tr)

}
