% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.node.R
\name{is.node}
\alias{is.node}
\title{Test if an object is a node}
\usage{
is.node(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
A logical.
}
\description{
The function \code{is.node} returns \code{TRUE} 
if \code{x} is a node, \code{FALSE} otherwise. 
A 'node' is just an \code{rtree} object with no subtrees.
}
