% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{is.root}
\alias{is.root}
\title{Test if a node is a root of a tree}
\usage{
is.root(.node, .tree)
}
\arguments{
\item{.node}{A node or node label to be tested.}

\item{.tree}{A tree}
}
\value{
A logical.
}
\description{
The function \code{is.root} returns \code{TRUE} 
if \code{.node} is a root of \code{tree}, \code{FALSE} otherwise.
}
