% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{is.rooted}
\alias{is.rooted}
\title{Test if a tree has a root}
\usage{
is.rooted(.tree)
}
\arguments{
\item{.tree}{A tree to be tested.}
}
\value{
A logical.
}
\description{
The function \code{is.rooted} returns \code{TRUE} 
if \code{.tree} is a rooted tree, \code{FALSE} otherwise.
}
