% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.tree.R
\name{is.tree}
\alias{is.tree}
\alias{is.rtree}
\title{Test if an object is a tree}
\usage{
is.tree(x)

is.rtree(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
A logical.
}
\description{
The function \code{is.tree} returns \code{TRUE} 
if \code{x} is a tree, \code{FALSE} otherwise.
}
\examples{
## FALSE
is.tree(empty_tree())

}
