% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeClass.R
\docType{class}
\name{Nodes}
\alias{Nodes}
\title{Nodes Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Nodes
}
\description{
Nodes class
}
\section{Fields}{

\describe{
\item{\code{labels}}{character. A vector of labels for each node. The labels must be part of 
the \code{tree}'s labels.}

\item{\code{tree}}{rtree. The tree where the nodes belong to.}

\item{\code{...}}{Additional parameters (currently not used).}
}}

\section{Methods}{

\describe{
  \item{get_labels()}{Get labels of the nodes}
  \item{get_tree()}{Get the tree attached to the nodes}
  \item{set_tree(tree)}{Set the tree}
  \item{keep(i)}{}
}
}

\seealso{
\code{\link[oak]{Tree}} and \code{\link[oak]{Node}} 
in this package.
}
\keyword{datasets}
