% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rtree}
\alias{print.rtree}
\title{Print trees}
\usage{
\method{print}{rtree}(x, at = NULL, level = 1L, ...)
}
\arguments{
\item{x}{A tree.}

\item{at}{character. If not \code{NULL}, the name of an attribute of the nodes of \code{x} 
to be printed next to the node labels.}

\item{level}{integer. This argument is used internally by the function, should not be used directly.}

\item{...}{Additional parameters (not used).}
}
\value{
\code{x} is return invisibly.
}
\description{
This function prints the labels of the nodes of a tree, 
and displays the connections between these nodes.
}
\seealso{
For examples using the \code{at} argument, 
see \code{\link[ginkgo]{tree_apply}}.
}
