% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{is.empty}
\alias{update}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bazar}{\code{\link[bazar]{is.empty}}}

  \item{stats}{\code{\link[stats]{update}}}
}}

