% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.R
\name{rev.rtree}
\alias{rev.rtree}
\title{Reverse a chain}
\usage{
\method{rev}{rtree}(x)
}
\arguments{
\item{x}{A chain to be reversed.}
}
\value{
The reversed chain.
}
\description{
This function reverses the order of the nodes 
in a chain.
}
\examples{
(tr0 = c_("Bob", "Carl", "Daniel"))
(rev(tr0))

}
