% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{root}
\alias{root}
\alias{root.rtree}
\alias{roots}
\alias{roots.rtree}
\title{Root(s) of a tree}
\usage{
root(.tree)

\method{root}{rtree}(.tree)

roots(.tree)

\method{roots}{rtree}(.tree)
}
\arguments{
\item{.tree}{A tree.}
}
\value{
\code{root} returns a node, the root of the tree if it exists;  
if \code{.tree} is not rooted, an error is thrown.  

\code{roots} returns a list of nodes, the roots of the tree.
}
\description{
Find the root of a tree 
(or the multiple roots for a non-rooted tree)
}
