% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take.R
\name{take_branch}
\alias{take_branch}
\alias{take_branch.rtree}
\alias{take_node}
\title{Take a branch of a tree}
\usage{
take_branch(.tree, .node)

\method{take_branch}{rtree}(.tree, .node)

take_node(.node, .tree)
}
\arguments{
\item{.tree}{A tree.}

\item{.node}{A node of \code{.tree}.}
}
\value{
A tree.
}
\description{
The function \code{take_branch} looks for the subtree of
\code{.tree} whose root is the node identified by \code{.node}.
}
