% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.rtree}
\alias{update.rtree}
\title{Update a tree with new subtrees}
\usage{
\method{update}{rtree}(object, subtrees = NULL, ...)
}
\arguments{
\item{object}{A tree to be updated.}

\item{subtrees}{A list of trees.}

\item{...}{Additional arguments to be passed as attributes to each 
node of the tree.}
}
\description{
BLA BLA
}
